Added 3 new methods to DataBaseConnection.java, DataBaseConnection.addTransaction, DataBaseConnection.addTransactionDetails, DataBaseConnection.getBarcode and DataBaseConnection.getTransactionID
Created new JPanel Class GrandTotal.java, which acts as the container for A JTextField to display Total Amount and its identifying JLabel

Added a FocusListener to InputPanel.java to shift focus to Product_Name JTextField when the Panel is in focus

Added New GrandTotal field in OutTable class and added the GrandTotal panel below the Table to display total. Changed the response of getting emitted Text from inputPanel to update the GrandTotal correspondingly

Modified Table.java. Modified Table.updateRow method to add the functionality to update quantity if same product is added instead of creating a new row to show the item