Basic Functionallity complete. DataBase I/O working perfectly
Updated ActionPerformed Interface by adding actionCommand method

Created new Interface CommitListener with two methods, returnListener and updateListener which couples CommitPanel and InventoryUpdate

Created new Interface CredentialListener which only has a single method to get credentials. it couples Login and POSMain

Created new Interface Row which have all the relevant methods for my TableRow Class

Created a final class ColoringConstants which have static fields of type Color which are used by different UI elements

Updated class GrandTotal to add the functionallity to increment the total

modified class InputPanel.java, changed the constrtuctor to accept the Button name to generify the Class. changed Font of button

Created Class SavePanel that holds two Jbuttons. this class is used in POSMain

Created Class Table.java, this class is the actual container that holds TableRow, this class is responsible to diplay the data and to acts a storage area for data while its being manipulated.

Modified class TableRow.java. this is the class that Table.java uses for its row.
overloadeded the constructor so that the class can be used in multiple ways.
TabelRow() is the default implementation of the class with only the Quantity Column updateable
TableRow(rowEnabled) is enhanced implementation that enables all columns thus allowing the class to used in the InventoryUpdate.table class
TableRow(rowEnabled, isHeader) is another enhaced implementation that formats the textFields such that they are used as Table header
Decreased height from 30 pixels to 25 pixels
TableRow.updateRow(String) works for everytype of TableRow enhancment
TableRow.updateRow(String[]) is still a work in progress. in the future updates it might be removed
Overloaded TableRow.incrementQuantity() and added TableRow.incrementQuantity(int) so increment by a specific quanitty
added method TableRow.saveData() which return a String[] with productname and quantity only.
added method TableRow.updateData() which returns a String[] with all Column data.
added method TableRow.reset() that sets every Column to an emoty String.
addded setter for the Field listener.

Created a new Class UpdatePanel that entends InputPanel.java. it add another textfield and a button to the existing class and adds cosmetic effects

Created class CommitPanel.java that holds two JButtons. it is used by the InventoryUpdate class to provide buttons to Commit wither Update or Return

Modified Class DataBaseConnection
Modified constructor to accept the username and password of the database thus i utilize the DataBase's own Login as a Login for the application
cleaned some redundant code and debug code
added Method DataBaseConnection.productNames() that return all the products that contain the characters of the query to be used as autocomplete data
updated method DataBaseConnection.searchQuery to add the barcode into the seach parameters
added method DataBaseConnection.addReturn() which adds single entry with current timestamp in the Returns Table
added method DataBaseConnection.getReturnID() which returns the last ReturnID
added method DataBaseConnection.addReturnDetails() which adds all the return items to the Return_Details Table.
added method DataBaseConnection.updateInventory() which updates the Inventory.

Created Controller Class InventoryUpdate.java that provides the main JFrame for the Inventory Update Application

Created Class Login which is a lightweight application that provides an interface to Input username and password and it just simply passes that information to DataBaseConnection

Created Main class that acts as experimental Run Config

Mpdified the POSMain to add all the basic functionality as needed